requirements.txt
# YOLO标注与训练自动化工具

# GUI框架
PyQt5>=5.15.0
Pillow>=9.0.0

# YOLO训练
ultralytics>=8.0.0

# ONNX转换
onnx>=1.14.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0
opencv-python>=4.5.0

# 报告生成
matplotlib>=3.5.0
seaborn>=0.11.0

# 工具库
PyYAML>=6.0
tqdm>=4.62.0
